moved DICE form classification_metrics.py to image_metrics.py
