+++ src/metrax/classification_metrics_test.py
@@ -47,10 +47,6 @@ SAMPLE_WEIGHTS = np.tile(
     (BATCHES, 1),
 ).astype(np.float32)
 
-DICE_ALL_ONES = (jnp.array([1, 1, 1, 1]), jnp.array([1, 1, 1, 1]))
-DICE_ALL_ZEROS = (jnp.array([0, 0, 0, 0]), jnp.array([0, 0, 0, 0]))
-DICE_NO_OVERLAP = (jnp.array([1, 1, 0, 0]), jnp.array([0, 0, 1, 1]))
-
 class ClassificationMetricsTest(parameterized.TestCase):
 
   def test_precision_empty(self):
@@ -81,13 +77,6 @@ class ClassificationMetricsTest(parameterized.TestCase):
     self.assertEqual(m.false_positives, jnp.array(0, jnp.float32))
     self.assertEqual(m.false_negatives, jnp.array(0, jnp.float32))
     self.assertEqual(m.num_thresholds, 0)
-    
-  def test_dice_empty(self):
-    """Tests the `empty` method of the `Dice` class."""
-    m = metrax.Dice.empty()
-    self.assertEqual(m.intersection, jnp.array(0, jnp.float32))
-    self.assertEqual(m.sum_true, jnp.array(0, jnp.float32))
-    self.assertEqual(m.sum_pred, jnp.array(0, jnp.float32))
     
   @parameterized.named_parameters(
       ('basic_f16', OUTPUT_LABELS, OUTPUT_PREDS_F16, SAMPLE_WEIGHTS),
@@ -274,36 +263,6 @@ class ClassificationMetricsTest(parameterized.TestCase):
         rtol=1e-2 if y_true.dtype in (jnp.float16, jnp.bfloat16) else 1e-7,
         atol=1e-2 if y_true.dtype in (jnp.float16, jnp.bfloat16) else 1e-7,
     )
-
-  @parameterized.named_parameters(
-      ('basic_f32', OUTPUT_LABELS, OUTPUT_PREDS_F32),
-      ('low_threshold', OUTPUT_LABELS, OUTPUT_PREDS_F32),
-      ('high_threshold', OUTPUT_LABELS, OUTPUT_PREDS_F32),
-      ('batch_size_one', OUTPUT_LABELS_BS1, OUTPUT_PREDS_BS1),
-      ('all_ones', *DICE_ALL_ONES),
-      ('all_zeros', *DICE_ALL_ZEROS),
-      ('no_overlap', *DICE_NO_OVERLAP),
-  )
-  def test_dice(self, y_true, y_pred):
-    """Test that Dice metric computes expected values."""
-    y_true = jnp.asarray(y_true, jnp.float32)
-    y_pred = jnp.asarray(y_pred, jnp.float32)
-
-    # Manually compute expected Dice
-    eps = 1e-7
-    intersection = jnp.sum(y_true * y_pred)
-    sum_pred     = jnp.sum(y_pred)
-    sum_true     = jnp.sum(y_true)
-    expected     = (2.0 * intersection) / (sum_pred + sum_true + eps)
-
-    # Compute using the metric class
-    metric = metrax.Dice.from_model_output(
-        predictions=y_pred,
-        labels=y_true
-    )
-    result = metric.compute()
-    
-    np.testing.assert_allclose(result, expected, rtol=1e-5, atol=1e-5)
 
 
 if __name__ == '__main__':
